pyproject.toml
[tool.poetry]
name = "stock-app"
version = "0.1.0"
description = ""
authors = ["alejovillores <avillores@fi.uba.ar>"]
readme = "README.md"
packages = [{include = "stock_app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
importlib-metadata = "^6.8.0"
importlib = "^1.0.4"
alembic = "^1.11.2"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
flake8 = "^6.1.0"
black = "^23.7.0"
sqlalchemy = "^2.0.20"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.poetry.scripts]
start = "stock_app.main:app"